+++ client/web/src/nav/UserNavItem.tsx
@@ -5,7 +5,6 @@ import { Shortcut } from '@slimsag/react-shortcuts'
 import classNames from 'classnames'
 // eslint-disable-next-line no-restricted-imports
 
-import { Toggle } from '@sourcegraph/branded/src/components/Toggle'
 import { useKeyboardShortcut } from '@sourcegraph/shared/src/keyboardShortcuts/useKeyboardShortcut'
 import { useCoreWorkflowImprovementsEnabled } from '@sourcegraph/shared/src/settings/useCoreWorkflowImprovementsEnabled'
 import { ThemeProps } from '@sourcegraph/shared/src/theme'
@@ -23,7 +22,6 @@ import {
     Select,
     Icon,
     Badge,
-    ProductStatusBadge,
 } from '@sourcegraph/wildcard'
 
 import { AuthenticatedUser } from '../auth'
@@ -170,17 +168,6 @@ export const UserNavItem: React.FunctionComponent<React.PropsWithChildren<UserNa
                                         </small>
                                     </div>
                                 )}
-                            </div>
-                            <div className="px-2 py-1">
-                                <div className="d-flex align-items-center justify-content-between">
-                                    <div className="mr-2">
-                                        Simple UI <ProductStatusBadge status="beta" className="ml-1" />
-                                    </div>
-                                    <Toggle
-                                        value={coreWorkflowImprovementsEnabled}
-                                        onToggle={setCoreWorkflowImprovementsEnabled}
-                                    />
-                                </div>
                             </div>
                             {!isOpenBetaEnabled && props.authenticatedUser.organizations.nodes.length > 0 && (
                                 <>